Create oAuth client for frontend service
This is work on #27, just the part that sets up and provides the details
to Kubernetes for the oAuth authentication tokens.

This PR:

* Generates an internal oAuth brand and client token
* Provides instructions on how to make it external
* Generates a `ConfigMap` with the env data that will need to be passed
  to the eventual `frontend` Pod container.
* Deploy said ConfigMap to the Services cluster.

I also rename "frontend-api" to "frontend", since the "api" part seemed
redundant.

Next should be Cloud Build the image and Deploy it as a `Deployment` and
`Service`.