Agones: Fix installation timeout.
Installing Agones to an Autopilot would sometimes fail on a variety of
health checks.

The answer is to use `tolerateFailuresUntilDeadline`, but that had to
include an upgrade to the latest Skaffold version (which we likely had
to do at some point anyway).

Was tricky, but at least now we have a template for upgrades.