Add a basic match function
* We use a MatchProfile per region so that the match function is
evaluated for each region and all open tickets

* In the match function, for a given region, we sort tickets by some
"score" that includes latency to that region and skill and assign
matches for that region. We assign an evaluator score based on the
aggregated score on the ticket.

* The evaluator score should let us assign ideal matches between the
regions, so in general the latency should be a big component of that
score.